requirements.txt
streamlit
google-genai
python-dotenv
fastapi
uvicorn
elevenlabs
